Got rid of warnings (and fixed several bugs along the way).
git-svn-id: https://hypertable.googlecode.com/svn/trunk@458 01261837-d133-0410-8dcb-6792eb736156

+++ src/cc/AsyncComm/Serialization.h
@@ -183,7 +183,7 @@ namespace Hypertable {
       (*remainingPtr) -= 4;
       (*bufPtr) += 4;
       // data
-      if (*remainingPtr < *lenPtr)
+      if (*remainingPtr < (size_t)*lenPtr)
 	return false;
       *dstPtr = (*bufPtr);
       (*remainingPtr) -= *lenPtr;
